rasgoql/requirements-dev.txt
google-auth-oauthlib
google-cloud-bigquery
jinja2
pandas
pre-commit
python-dotenv
pyyaml
rasgotransforms~=1.3
requests
snowflake-connector-python
snowflake-connector-python[pandas]
redshift_connector
pytest
